Implement nested routing for demo routes
Refactors the structure of the Unauthenticated app component tree to allow for nested routing and demo app styling that differs from the login page, which is also in the unauthenticated app tree.

This was a difficult solution to figure out due to React Router's documentation/changes across more recent versions. At the time of implementation this app was using React Router 5. This Stack Overflow answer was particularly helpful in solving this:

https://stackoverflow.com/questions/56711663/react-router-v5-0-nested-routes