Add unauthenticated predictions client for demo app
Adds a new API client, DemoAPIPredictionsClient, which is similar to the APIPredictionsClient implemented earlier, but which doesn't require an authoritzation token to initialize. This is to allow users demoing the app to try submitting a prediction for a recent show.

Breaks shared fetch call to load prediction categories into a base class that both DemoAPIPredictionsClient and APIPredictionsClient extend from.

DemoAPIPredictionsClient will eventually implement its own submission to a new unauthenticated demo prediction endpoint which will score the prediction in real time and display the result.

This is the first stage in an attempt to reuse all of the complex prediction form logic in src/components/predictions/predictionForm.tsx by decoupling it from the authenticated (i.e. registered user) use case so it can be used for the demo use case.