Refactor Predictions form for demo and authed versions
Overhauls the PredictionForm component that was previously coupled to submitting predictions for authenticated users to spin off two flavors of form that inherit the same complex logic - one with an injected permission submission client that requires an authentication token and one that does not.

The unauthenticated predicitons client will submit predictions to a new endpoint that will score an unauthenticated user's retroactive predictions for the latest concert played.

Blocks in where the demo predictions form will live once the backend endpoint for demo predictions is implemented, in src/components/UnauthenticatedApp/Demo/latestConcertPredictionPage.tsx